server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.111.0
uvicorn>=0.30.0
openai>=1.35.0